Beta release of registration process.
Some cosmetic changes are still necessary to further improve this
project. Most of the redirections have been reallocated to the
WebServer class. Still missing is code in the lib/* folder.

Mail is not guaranteed to work as it has not been tested in an
environment capable of sending mail. Therefore, the entire process
of creating a new account has not yet been tested.

Space for the EULA has been designated, but it has not yet been
created.