PUBLIC: Speed ups and setup fixes.
This release includes major changes to the way the entire site is
organized, taking advantages of virtual hosts to make sure that the
public side is hosted from a more 'sterile' (sub)domain.

In addition, some left over Makefile rules and apache.conf directives
were added to the default configuration. Note also some changes to
conf.default.php. In short a full review of the site is required when
updating to this revision. Prepare accordingly.