Major changes to database: merge finish and score tables.
Essentially, there is really no need to have finish and score
as two separate tables. This commit merges those two into one
finish table, significantly reducing the number of calls to
the database and thus speed-up the overall performance.

Note that the Score class is still active and this is an alpha
release. Further testing is necessary.