Merged 'race' and 'race_num' tables into one table.
This required changes to the DB, of course, but also changes to the
API, particularly the method 'addRace' has been deprecated in favor
of 'setRace' which should be slightly more trustworthy. This change
should be completely transparent to the rest of the code, but further
testing would not hurt.

Note that the race_num table is not removed yet, just in case.