Fix bug with PHP 5.2.6 with function names serving as constructors.
It turns out that prior to 5.3.0, a class' method which had the same name
as the class itself, served as a possible constructor for that class.
This is a problem, of course, if that constructor is already defined
appropriately using __construct.

Note that, due to the case-insensitivity of both class and function names
in PHP, it did not matter that the class name was SendMessage and the
function sendMessage.