Sending messages as an offline process.
In this rewrite of the message sending abilities, the web interface
merely queues a message to be sent into a new table called 'outbox'.
The actual message is then sent by a scheduled run of the script
ProcessOutbox.php, set to run once every hour, 27 minutes past the
hour according to crontab.default.

This makes the process better able to withstand the load caused by
ever more users in the system.

Note that this commit includes database changes.