Fix AllAmerican report to include sailor participation for all regattas.
Whereas before the program was including merited sailor's finishes for
only that regatta in which they so earned that automatic inclusion into
the report, this fix now ascertains that participation in all regattas
being included in All American is included in the report.

This once again makes the All America report generation very slow and
memory intensive. In addition, it *must* be made clear, from a code point
of view, that the fillHTML and process methods actually work in tandem,
both being involved in the process of generating the code.