Preliminary release of Head-to-Head Compare Sailors.
While this release showcases vast improvements by using the public
information for the regattas (including Dt_Rp), instead of the more
resource-intensive, older method (Regatta objects), it also uncovered a
bug in the data which will need to be fixed.

At some point, as yet undetermined, finalizing a regatta resulted in the
unscored races being automagically removed from the database. This needs
to happen again. We also need to look at the process of finalizing a
regatta and the kind of Update_Request that it queues, as it might be
more "intense" than a DETAILS change.