Migrated FItem to new HtmlLib environment.
This commit introduced the need to properly address in the future the
conversion from 'addChild' to 'add' and 'addAttr' to just 'set'.

It should be technically possible to do that next, for instance, so long
as the new methods are replicated on XmlLibrary's GenericElement. As of
right now, the old methods are being temporarily replicated in XmlLib,
but a good idea at this stage would be to "turn the switch".