Whew! Migrated Para to XP.
(No, Windows 7 is not on the radar.)

Checked for PHP syntax errors; but have not thoroughly checked the
logic. However, all the paragraphs were "intelligently" migrated to fully
use the feature set provided by the superior XP class.

This should result in vast memory improvements, reduction of LOC's, and
overall health for the subsequent programmer.

Also, more ROI.