Mature XSelect migration for EnterPenaltyPane and DetailsPane.
It seems obvious that the XSelect which comes with HtmlLib might use some
upgrading to match the facilities provided by XmlLibrary. For instance,
it is very common and often useful to create a drop down list directly
from an associative array of elements.

To this end, I have MODIFIED the HtmlLib to add static function
::fromArray which will generate one such object.

In addition, given that drop down lists where one can select multiple
elements are sufficiently different in nature from drop down lists
themselves, I propose a new subclass of XSelect called XSelectM precisely
for this purpose.

Note that, unlike the previous commit, I do believe that XSelects should
retain their current argument ordering with attrs before items, as it is
more of a "block-level" element and would make it consistent with the
other elements in the library.

With the addition of ::fromArray, this old order should be less of a
problem.