Removed nonprint class from RpEnterPane ports.
Though this was an interesting idea, we need to further explore it down
hte line. As it is, only RpEnterPane used the 'nonprint' attribute for
Ports. It might be useful, down the line, to create a new Port derivative
class to take care of these things, as the current mechanism does not
work.