Migration of RpManager static function get[Unregistered]Sailor[s] to DB.
It occurs to me now that these functions, whose first argument is the
school object, should really go in the school class itself, rather than
generically part of DB. However, due to the complexity introduced by that
rewrite, this temporary migration will have to do.

Next, we cover coaches, which are like sailors.