Migrated and fixed Host/Scorer API (untested).
It used to be that a 'host' was a TS account object. Then, the
host-school was the school for that account. At some point a while
back, it was decided that an account should be allowed to score a
regatta without being a host of the event. As such, the decision was
made to split the functionality of host and scorer so that the table
host_school linked regattas and host schools (many to many) and the
table scorer linked regattas and accounts (many to many).

Unfortunately, much of the confusion regarding the new roles remained
in the code in order to facilitate the new feature release at the
time. One such confusion is that the function Regatta::getHosts
returned Account objects, instead of the more natural School
objects.

The role of this commit is to at once migrate Host_School to use the
DBM library, and fix Regatta::getHosts to return school objects.

As it turns out, this distinction had been already made for the case
of Dt_Regatta in the public API.

The Host class has been refactored as Scorer, since that is what it
is.