Migration of Season constructor to Season::forDate.
This process has highlighted the need for a newly implemented static
function Season::getRegattasInSeasons which returns all the regattas
that are in all the seasons provided. This should replace all instance
of array_merge() which are both (a) illegal for DBDelegates and (b)
inefficient to begin with.

Also removed Season.php object, and cleaned up repeated instance of
Season object in DB.php (it has been created and commented out some
revisions ago).