Some migration of Regatta into RegattaSummary.
The idea here is to make RegattaSummary be the same as Regatta
currently is. That way, the migration from old method to DBM will be
seamless and less painful.

In the process, we discovered we need better support for Singlehanded
Teams, which are still not migrated over to the new mechanism. This
might have to wait until such a time as the Regatta migration is
complete so that Team's 'regatta' attribute can point to the DBM
regatta object.