Cache Regatta creation by moving constructor to DB::getRegatta.
The idea here is to provide a temporary mechanism (eventually to be
replaced by the natural caching mechanism of DBM), so that at least
some derivative objects like Team can have access to their regatta,
and not just to their regatta's ID.

This allows us to have some of the benefits of DBM before actually
having DBM.

For now, DB::getRegatta forwards Regatta's constructor's exceptions
along. A possible future fix is to eliminate this exception and return
NULL, as usual, for non-existing or invalid ID.

Armed with this changeset, we can now tackle the problem of
Singlehanded teams.