Changed seasons so that their ID is their short name.
This kind of DB change *should not* cause problems at the time of
updating. But watch for dt_regatta rows which now have null for their
season. These need to be addressed.

A wonderful consequence of this change is that now the last of the
non-trivial functions in Preferences has been migrated onto the DB
library. We are almost good to fully remove Preferences.php!