Migrate the last use of Preferences::query() to DBM!
With this commit, we have completely deprecated (and removed) the
Preferences class.

Although we have now removed all obvious traces of a database
background by removing all MySQL-related calls, the UpdateSailors
script has suffered a bit in performance.

Whereas one query used to take care of both insert or updating (by
using the MySQL-specific 'on duplicate key' clause), the new system
requires checking for the existence of the ICSA sailor before calling
DB::set() with either 'true' or 'false' as the second argument. This
means that there is an extra SQL call FOR EACH sailor.

Technically, the old method was not actually optimized; as it created
one query per sailor instead of one query for many sailors.
Nonetheless, the new system does not allow for multiple querying of
this nature. Just something to keep in mind.