Intermediate commit before addressing Division property bug.
The problem we are experiencing is that the 'division' property of Races
(and of other objects as well) requires it to be a Division object,
which the DBM is insisting on serializing to the database (instead of
casting as a String).

There are two ways to address this problem. Perhaps the cleanest is to
affect the DBM library, by slightly modifying the serialization
process to honor the type from DBObject::db_type. This way, the
serialization process can force the casting of the Division object as
a String (which would be correct) at the time of serialization, and
the Race subclass can override &__get to make sure that a Division
object is always returned.