NewRegattaPane is not just Soterified, but respect team racing.
This includes numerous changes that are only partially related to the
end-goal of including some Team Racing. For example, AbstractUserPane
also behaves like AbstractPane in the procesPOST wrapper.

In addition, NewRegattaPane NO LONGER requires creation of races and
divisions because, in the case of team racing, divisions do not really
make sense (actually there are all division A, shhh).

As a result of this change, it is imperative that the next pane the users
see is the one where they create races. Then, the one where they create
teams. All three are necessary, recall.

Finally, NewRegattaPane will not allow the creation of a regatta in a
season that does not yet exist.

Lots to do, still, but this is good progress, I'd say.