Feature: TeamRacesPane specifically for team racing regattas.
Team racing regattas require knowledge of which pair of teams participate
in which race. As such, adding races to a team race regatta requires the
user pick the teams (and the boat) for each race before the race gets
added.

In addition, it is also possible to edit the teams that participate in
any given race, but only if that race has not yet been scored. If it has,
then users can only edit the boat for that race.

Finally, the current race-editing input requires users to check the box
for the race they want to edit (I really like tabular display for this
data and know of no other way to do it, really). This can be vastly
improved with the use of Javascript to do the checking automatically.

Unfortunately, the Javascript is not yet done.