Finished fixes to SendMessage pane and ProcessOutbox.
New information layout is more robust and future proof. Still to come is
the ability to adderss specific user(s). Although the necessary constant
has been added to the Outbox class.

In addition, we should implement multiple verbose flags for ProcessOutbox
to see more of what is happening, since the script can take a long time
to complete.

Finally, Message::$active is > 0 for active and NULL for 'deleted'. This
is the reverse of the usual convention established. I propose changing
this to the more conventional Message::$inactive, with NULL as default
value.