Folded preference panes into AbstractPrefsPane (new) and index.php
This continues are quest to simplify the class hierarchy and improve code
reuse. With this fix, we no longer need a separate prefs.php (it was
removed). Instead, the new AbstractPrefsPane does the extra check for
valid school, etc.

Please note that all the preferences pane will have to be soterified.