Release of the race's "round" feature.
The idea is that team races are grouped into rounds, which allows for
easy creation of the races (the user need only select the teams in the
round, and the round-robin races are correctly generated); and makes
feasible the possibility of the grid score layout (introduced in
tscore/ScoresGridDialog, to be released later).

For consistency, we need to give the new race's round attribute meaning
for non-team racing regattas. In those cases, the round number stands for
the "day" in which the race was scored. Thus, we can group Saturday races
differently from Sunday races.

Naturally, the day the race was scored depends can be backwards-derived
from the timestamps of the finishes for that race. Indeed,
changes.history.sql uses this logic to update the round attribute for all
past regattas. In the future, however, a race will receive its "round"
only the first time it is scored. Subsequent re-scoring of a race will
not affect its round number.

This commit does not include all the goodness of this feature set. Much
work is needed on TeamRacesPane to accomodate for the new paradigm. We
shall focus our attention on the grid layout next, since that it is a
critical component of team racing.