Release of ScoresGridLayout and major bug fixes.
Yes, the ScoresGridLayout is not very pretty, and its own CSS must yet be
developed, but at least the scoring is correctly carried out now (that is
to say that non-participating teams are added with DNS in the backend);
and the display works.

The problem is that the CSS for dialogs is in dire need of a major
overhaul.

I also fixed issues with the ICSA*Scorer's. These need to be checked.

ICSACombinedScorer is used for both TEAM and COMBINED scoring now, as
well. More work is merited. Particularly, we notice that entering
finishes in combined scoring is surprisingly slow, in no doubt due to the
number of DNS being scored all the same way (this amounts to a great deal
of update requests: hence the slowness).

Fixing this issue is, naturally, non-trivial. As such, we will put it off
as a wishlist item while we work on the more important kinks.