Greater security by hashing passwords with salts.
The goal here is to vastly increase the security of TechScore accounts
should there ever be a breach and the database be compromised. The
safeguards put in place match those of the latest, best security
practices, and include salting all passwords with a two-part salt: one
that is unique to each user (the username) and the other that is unique
to the application and stored separately from the database.

Note that this has been implemented without the need for users to reset
their passwords. It is important that the hash NOT be lost. If it is,
then all passwords would have to be changed.