Update404 is now a full-fletched AbstractScript.
It is much more user friendly. In addition, this is the first script to
make use of AbstractScript's writeXml and writeFile.

The arguments to both of these is now merely a relative filepath, without
assuming the existence of the 'html' directory. These details will be
abstracted away to the forthcoming serializers.