UpdateBurgee has been AbstractScript-ified. Interface changed.
AbstractScript received a new method to recursively remove a directory or
file (AbstractScript::remove).

UpdateBurgee's interface changed. The 'update' method has been refactored
as 'run' and the previous method by the same name obviously removed.
Thus, it is not possible to run UpdateBurgee in order to update all the
schools, as was the case before.

Because of this, the update to the burgees every 8-hours was removed from
the crontab as well.

The idea is that we queue school updates similar to how regatta updates
are currently queued in pub_update_request. Then, the UpdateDaemon can
make the associated changes whenever they actually happen in the admin
web interface.

We work on this concept next.