Regatta types are now separate database objects, instead of enumeration.
This means that there can be any number of regatta types, although at
least one must exist (note that the old types have been migrated for
backwards compatibility). The regatta types may also be ranked, and they
are fetched by this rank.

Still pending is the ability to modify these types and their rank online,
as well as a method by which they can be inactivated, etc.

Also pending is a method by which regattas can be sorted using the
regatta types' ranks.

Among the changes:

Deprecated Conf::$REGATTA_TYPE_BLACKLIST as this trait is saved with the
type itself in the attribute 'inactive'.

Deprecated Regatta::getTypes, now only used by the deprecated RegattaIO
script.

Deprecated Regatta::TYPE_* constants, now only used by currently
deprecated AllAmerican.php script, to determine which regattas
automatically fit the bill for inclusion in the report. In the upcoming
rewrite of this functionality, these calls will be replaced.