For combined division, only display the combined rank table.
This should fulfill the request from Matt Lindblad regarding the display
of the sailor table (known as the "Division dialog") for combined events,
such that a team's A division also sails against that team's B division
and is therefore ranked accordingly.

Since the files that are serialized are different for each regatta type,
this update will actively remove pages that should not exist for a given
regatta scoring type. Thus, it should be possible to update a previous
event by running:

php lib/scripts/UpdateRegatta <id> score

It should be possible to have the Division Dialog and the new Combined
Division Dialog to only use the information from the database, rather
than using the ranker themselves every time.