Store team score in cache tables (dt_*), automatically when scoring.
This deprecates UpdateRegatta::syncTeams (which has been removed). The
next step is to use this information directly when creating the tables
(rather than regenerating this information).

However, this preliminary update means that now even personal regattas
get stored in the dt_regatta table. But this information *must not* be
published. So next, we treat the program with an immunity to personal
regattas existing the dt_regatta table.