Store sailor's rank and explanation for races sailed in dt_rp.
This is a major step in data collection, but comes at a processing cost:
the need to possibly re-rank partial regattas whenever either an RP is
affected, or a new race is entered.

To appease this loss of efficiency, the RP syncing should still be
performed on demand by UpdateRegatta, but more importantly, it might be
necessary to only rectify one school's RP information at a time in the
UpdateRegatta::syncRP method. We tackle these in turn.

For data integration, it might be useful to run a Regatta::doScore on all
regattas, as well as an UpdateRegatta::syncRP().