New update activity: season change.
This changeset allows for a possible, although unlikely, event that a
regatta's start_time change pushes it to a different season
altogether. It is a stepping stone towards being able to manage
seasons from a web interface.

Since changes to season require deletion of previous regattas, we also
introduced a new request activity called URL. However, we have not
implemented any UpdateRegatta or Daemon logic for addressing this
activity.