Do not allow switching regatta scoring types to/from Team.
The reason is that the two scoring mechanisms are so disparate, that
there is no graceful way to switch from one to the other.

This also allows us to simplify the finish table so that we no longer
need to record DNSes for the teams that did not participate; although
this change is not implemented in this commit. This will have the benefit
of reducing the database, and allowing the use of DNSes for actual
penalties, thereby simplifying the code as well.