TR: Better support for tied teams.
As with fleet racing, Regatta::setRanks is called whenever a finish is
altered. This triggers the tiebreaking mechanism ICSATeamRanker which
will leave teams as tied by default.

The RankTeamsPane can be used to override the output of ICSATeamRanker,
and should be used for the purposes of breaking ties. However, any
intermediate tie breaking will have to be re-performed whenever a new
score is entered, even if the newly scored race does not involve any of
the tied teams.

Explanation for tiebreaker is not touched by ICSATeamRanker.

The ranking dialogs are faster now because they do not re-rank the
teams. In addition, ICSATeamRanker no longer honors the saved value of
the team's dt_rank attribute.