TR: Replace one or more teams within a round.
In doing so, replace the finishes associated with each replaced team, as
well as the rotations, and delete the RP information of the replaced team
as well, in order to ensure database correctness.

Still to investigate is (a) what happens when there are no possible teams
to replace, and (b) what happens with races that are carried over?