TR: Include next races to be sailed in rotations pane.
This is only made available to the private site, until approved for use
in the public pages.

The next races to be sailed prints the "next" three flights to be
sailed, including the one current under way. Races mid-flight that have
been sailed are so delineated with a 50% opacity in CSS.

No "sailing next" table is printed if the races to be sailed are within
the first three flights of the entire regatta, or if all races have been
sailed already.

This display does NOT jump from one round to the next. That is, if only
two flights are left in one round, it does not print those two rounds
PLUS the first of the next round.