(Intermediate) Split enter finishes into two steps for fleet racing.
The goal is to minimize the times that the wrong race is automatically
chosen, leading to the wrong finishes being entered.

With the "Enter finishes" pane split in two, there is an opportunity to
improve the race-selection mechanism, perhaps by using a drop-down
instead of a plain text box.

We attempt this next.