Use md5sum to determine if codebase has changed.
As explained in src/README.md5sum, this allows a mechanism by which the
Daemons can self-destruct if the codebase changes. This is important so
as not to use stale data from memory.

The process entails creating the file src/md5sum (which is now ignored).
This file is created by 'bin/Make.php md5sum', as performed by Makefile
in response to changes to lib/ directory.

Meanwhile, the lib/scripts/Daemon.php will now check the contents of the
md5sum file whenever it "wakes up", and terminate with an exception if
the md5sum no longer matches the one in the file.

Presumably, after this update, the Daemons need not be restarted manually
every time there's a change to the codebase.