Use regatta URL cache to determine what pages exist and may be erased.
This is a substantial change to the logic in UpdateRegatta which, for
one, greatly simplifies the script, but also provides a method by which a
local copy (under the "html" directory) is not needed. This is a much
more robust architecture which was long overdue.

The change has been tested fairly thoroughly with standard, singlehanded,
combined, and team racing regattas.

Other fixes:

  - Deprecated Regatta::getCombinedRaces, which was not being used.