Major: Removed $sync from UpdateRegatta rules.
Under "normal" operation, the Daemon should NOT be responsible for
setting the data. Instead, it should be utilized for serializing pages by
fetching data only. This should help to prevent locks on tables which
cause long delays.

The $sync_rp rule should also be deprecated soon.

An artificial 'sync' target should be added to UpdateRegatta.php, in case
the information needs to be artificially synced (a la
bin/GenerateData.php).