Adding teams observes new rules about team names.
No longer will team names be cycled through the entire list of
preferred team names. Instead, the first preferred team name is always
chosen (or the school nick name, if none exist). New teams from a
school will use this team name plus a suffix, if necessary.

Suffixes are based on all the other teams from that school in the
regatta. If no team exists with the same base name as the preferred
name, then the suffix is empty. Otherwise, the suffix is the last of
the previous suffixes plus one.

At the same time, the first team's name also gets changed to have a
suffix of 1 if multiple teams exist. Visually:

Only one team:   Engineers

A new team added:
  - Enginners  ->  Engineers 1
  - [new team] ->  Engineers 2