Remove all team name preferences with a numerical suffix.
These are invalid team names, and there seems to be a slew of coaches
that did not get the message when editing the setting the team name
preferences.

At any rate, the new team-name-choosing scheme expressly forbids the
use of numerical suffixes for team names as these are used by the
internal numbering system.

In the process of doing so, it is possible that some schools may be
left without a name altogether.

We have adapted the language in the "Team name" preferences pane to
specify that the team names chosen are actually squad names.

Also fixed a grammar bug when a team does not have any registered team
name preferences.