Set global configuration allowed to super admins only.
Migrated the TS_FROM_MAIL setting from Conf::$TS_FROM_MAIL to the
database via STN::TS_FROM_MAIL. Added a new functionality to STN class to
provide default values for certain constants if none exist in the
database. This is used internally by DB::g() to make sure that some value
exists for certain parameters that require it.

It is important to set the value of TS_FROM_MAIL after this change.