Provide script to fix existing team names.
In the process, made EditTeamsPane a child of AbstractTeamPane, although
it does not yet use its functionality.

In addition, AbstractTeamPane::fixTeamNames is now public instead of
protected, and it returns the list of changed team names.

Finally, fixed a bug in AbstractTeamPane::fixTeamNames caused by team
names with a slash (/) in them confusing the RegExp delimiters.