Migrate e-mail template for pending regattas to new database scheme.
For backwards compatibility, use the following message (no e-mails will
be sent if message is missing):

<?php
/**
 * Migrate the e-mail message templates
 *
 * @author Dayan Paez
 * @created 2013-11-26
 */

require_once(dirname(__DIR__) . '/lib/conf.php');

DB::s(STN::MAIL_UNFINALIZED_REMINDER,
      "Dear {FIRST_NAME},

You are receiving this message because one or more of your regattas are not yet finalized. All official regattas *must* be finalized, and all RP information must be accounted for, in order to be included in reports and on the website.

Below is a list of regattas that need your attention, as well as an indication of what needs to be addressed:

{BODY}

Please take a minute to log in using the links above.

Thank you for your time,
--
Techscore Administration");

?>