Fix bug in Regatta::getScoredRacesForTeam() for team racing.
The bug failed to consider the first argument, the division, in the
resulting query for team racing events. As a result, the cached value for
RP completeness (Team->dt_complete_rp) may be off and should be
recalculated.

For this purpose, we have included the new bin/CacheRpComplete.php, which
takes one argument, the ID of the regatta.