Do not use school's names in regexp when calculating team names.
This is a classic case of mixing code and content and led to a bug when
the original team name included a slash (/). Technically, a check for
this kind of condition was in place later on in the code. However, the
true solution is to do away with embedding user input in a regular
expression.