Disable submit button after click when entering finishes.
This is an attempt to put a stop to Exceptions being thrown on busy days
when the user seems to make a follow-up request to submit finishes due to
the original request taking a long time. As a result, the unique key in
the finish table (race-team) detects a duplicate.

Although not a full fix, this will help us diagnose this problem further.