Track when sailor/seasons were active for a particular season.
Although this information is not yet used, this changeset does
introduce tables and functionality to track when a sailor or school
was active for a particular season. This will allow us to tailor the
interface to only show schools/sailors that were active for the season
of the given regatta, which is specially important for installations
that exhibit a lot of turnover.

The new sailor_season and school_season tables give us further
flexibility, by also tracking the date in which a sailor or school
were "active". In a slightly convoluted way, this will allow us to
reconstruct the daily sync data provided by the upstream service.
Since the database can be synced through the admin interface, we added
a special flag (--log) to perform this logging. The cron job has been
altered to include this flag.