One method of letting users change e-mail address...
This requires a number of changes to the core system which may not be
altogether necessary in the simplest of cases. We propose a new
procedure, different from this one, wherein the token for validity is
e-mailed and maintained in the session, rather than the database. This
also has the effect of not requiring changes to the schema.