Allow for different mail sending strategies; include SMTP
Add support for different methods of actually sending e-mail, with the
built-in PHP mail() function as default (set in conf.php). Include an
SMTP strategy as well, leveraging some components of PHPMailer project
(the minimum required to get this functionality going).

There are some caveats associated with using the included SMTP strategy:
first, one must also include the appropriate settings for that strategy
in conf.local.php. We opted not to use the DB to store these settings
because mail is used as a debugging tool and reliability is
fundamental. In addition, the choice of PHPMailer means some loss of
flexibility when sending "in-memory" attachments (this is a stock
PHPMailer installation which does not provide this functionality).
Finally, bear in mind that SMTP will indeed send messages, so ascertain
that the "Divert user" setting, if any, is properly set.

  - Ignore cache/* files.
  - Namespace for sending e-mails.