Tackle the race condition when updating attendees
This is being addressed via two channels: by changing the nature of
DB::insertAll() to "INSERT IGNORE"; and by disabling submit buttons on
page unload.

1. Add "ignore" flag to insert statement

This flag allows for a work around race conditions that may arise when
transactions involve potential deletes and inserts in a table containing
unique indices.

It is being used as part of an effort to fix errors seen when updating
RP information regarding either duplicate violations of the unique
team-sailor index (when adding one that's already there), or missing
foreign key constraints (when assuming a record exists that was deleted
by a different transaction).

2. Disable submit buttons on page unload

This is intended to eliminate instances of double-submits which may be
at the root cause of race conditions, especially when the page load
takes some time.