Introduce FleetRotation and new RotationPane
This represents a new paradigm for both creating and storing information
about a fleet regatta's rotation settings. The FleetRotation object is
the first to be fully namespaced and inherit from the newly introduced
AbstractObject, a descendant of DBObject that provides four auditing
columns:

  - created_by
  - created_on
  - last_updated_by
  - last_updated_on

To avoid confusion, the former "Rotation" class was refactored as
"RotationManager", consistent with "RpManager". Furthermore, the class
TeamRotation, formerly used only for team racing regattas, has been
repurposed as SailsList, since it is being used in the new
FleetRotation. A convenience method, getSailsObjects(), returns a list
of Sail entries encompassing the sail number and color.

The new RotationPane strives for unit-testable components and a full
separation of concerns. To that end, a number of helper classes have
been created to distribute the computational load:

  - FleetRotationParser: handles translation to and from args
  - SailsTable: encompasses the list of sails shown to the user
  - SailsRotators: to handle different rotation types
  - RacesRotators: to handle different rotation styles
  - RotationCreators: to leverage the above in assigning sails
     - FrannyFleetRotationCreator: specifically for "Franny" rotations
     - CombinedFleetRotationCreator: for combined-scoring regattas
     - StandardFleetRotationCreator: for the rest
  - FleetRotationValidators: suite of small, Soter-like components that
    verify the integrity of the FleetRotation attributes.

Unlike SailsPane, kept for auditing and evaluation purposes, the new
RotationPane uses FleetRotation constants for the different rotation
types and styles. In another notable departure, RotationPane is meant to
be used to create the rotation for the entire regatta, so the "Races"
parameter has been removed: the entire fleet regatta's rotation is (re)
created in one go.

A dedicated DeleteRotationPane has been added for deleting rotations.
This has the effect of simplifying the RotationPane, while separating
concerns.

The "No rotation" option has been re-labeled as "Same boat throughout",
to avoid linguistic confusion.

The list of arguments needed to create a rotation have been evened out
across the two steps: the rotation type, style, division order, and
races per set are bundled in the first step, while only the list of
sails are shown in the second. Note that the "Order sails" option has
also been removed.

To handle the fact that "Same boat throughout" rotations do not require
a "Races per set" argument, a standalone Javascript has been deployed.

A convenience method was introduced, Division::listOfSize(), which
should be used throughout Techscore whenever a number of divisions is
required.

Unit and integration testing is a critical component of this changeset.
All introduced classes (those in the "rotation" namespace) are fully
covered. To facilitate future integration testing, convenience methods
have been added to tst/integration/AbstractTester to facilitate XML
searching for components. In addition, POST operations are now fully
supported, including the use of the appropriate CSRF token.

Other noteworthy changes include:

  - Properly handle fatal errors

Move "division order" to Step 1 of RotationPane

  - Introduce ui/SortableTable and use it in RotationPane
  - Hide division order when style is "copy"
  - Add integration test for step 2 of RotationPane
  - Better FAQ + ProgressDiv
  - Recall parameters of last saved rotation

req https://trello.com/c/xUTvcevJ/212-track-the-rotation-options-per-regatta