Introduce AutoFinalize script
This is a script which will automatically finalize a regatta, and
depending on settings, assess a missing RP penalty to teams with
incomplete data. In addition, an e-mail is sent (aggregated by account)
for all the affected regattas and teams, according to the template, if
any.

  - Add admin setting parameters for auto-finalize
  - Introduce DB::getPastRegattasThatEndedAfter
  - Add global conf "ALLOW_AUTO_FINALIZE" setting.
  - Introduce AUTO_FINALIZE_ENABLED setting
  - Introduce StnCheckbox and use it in GlobalSettings
  - Add AutoFinalizePane
  - Make GlobalSettings namespace savvy
  - Turn AbstractSuperUserPane namespace savvy
  - Allow overriding regattas in AutoFinalize

req https://trello.com/c/V14jkPEY/207-automatically-finalize-regattas