Plug hole when counting RP as complete for a team
The bug, which is very subtle, arises because of the way that RPs are
counted when checking for completenes. To minimize database calls,
Techscore would count that the number of entries in the RP table (that
is combination of race-team-role-attendee) was greater than or equal to
the sum of the minimum number of people allowed in the boats.

A way to illustrate how to circumvent this method is to consider a
single-division regatta with 5 races using a boat that allows between 2
and 3 crews (i.e. a boat that requires between 3 and 4 people including
the skipper per race).

If I enter a full boat for the first four races, then Techscore would
count a total of (4 races * 4 people/race) 16 entries. This number
exceeds the expected minimum of (5 races * 3 required people/race) 15
entries, thus tricking Techscore into believing that the RP entry was
complete for that team.

However, this is not the case since the last race has no entries at all!

A unit test was added to check precisely for this case.

req https://trello.com/c/08bcdTyC/217-plug-hole-when-counting-rp-as-complete-for-a-team