Introduce the EnterFinishesWidget
The widget and subsequent new EnterFinishPane provide for an improved
user experience by replacing a confusing entry method with a more
intuitive one: the list of possible sails is separated from the pool of
entry; logical attempt is made to fit all the available sails in one
page fold; better visual feedback for both valid and error conditions.

Furthermore, the EnterFinishesWidget provides an abstraction that
greatly reduces the maintenance cost of the EnterFinishPane and its
much-simplified extension, TeamEnterFinishPane. Among the improvements
on this front is a pared down set of expectations when posting finishes
to the page, which does not care for the rotation. Instead, entering
finishes via the rotation is merely a visual aid to the user. In
addition, multiple races can actually be submitted at once, making the
page more generic. The result is a drastically simpler EnterFinishPane.

Unit tests are added for EnterFinishesWidget. Integration tests have
been added for standard and combined fleet racing for EnterFinishPane.

req https://trello.com/c/361hlYRY/197-scorers-score-first-race-s-of-regattas-incorrectly-often