Fix "Adding teams when only one and scored regattas" bug
As reported by e-mail to the user that found the bug:

By design, Techscore will not allow you to enter finishes if there's
only one team in a regatta, as this would not make any sense. As a
result, it assumes that, if there aren't at least two teams
participating, then there must not be any scored races either.

There's one edge case that isn't properly covered: scorer adds
multiple teams, enters finishes for one or more races, and then
removes all but one of the teams. In doing so, the scorer has left the
system in a state it assumes is not possible: one team with finishes
entered.

An integration test has been added to verify proper behavior. The
solution is to not rely on variables derived for UI presentation (such
as "has_score"), but to calculate based on actual data.